idea: future output management
in the current implementation Output property of the struct Property is
char*, while this satisfies the purpose, the code could be made more
extendable by having Output of type OutputMap pointer. this way, each
module can have varying length of output generated, stored in a way that
can be computed later.

implementing this change at the moment is at a lower priority than
getting the system work